Use cast package for castiing types, in all places
