+++ src/api/config_test.go
@@ -1,0 +1,7 @@
+package api
+
+import "testing"
+
+func TestConfigApiCalls(t *testing.T) {
+
+}
