+++ src/api/row_test.go
@@ -1,0 +1,7 @@
+package api
+
+import "testing"
+
+func TestConfigRowApiCalls(t *testing.T) {
+
+}
